requirements.txt
# Django AI Project - Requirements
# Install in project directory or use main requirements.txt

Django==4.2.7
openai==1.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0

